Add label type to ID mapping for various biological terms and gene ontology
- Introduced a comprehensive mapping of label types to their corresponding IDs, including terms from the Fly anatomy ontology (FBbt), Gene Ontology (GO), and Sequence Ontology (SO).
- This mapping enhances the utility of the application by providing standardized identifiers for various biological entities, facilitating better data integration and analysis.
- Exported the new mapping as part of the module for accessibility in other components.
